requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
asyncio==3.4.3
attrs==23.2.0
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2.7
django-debug-toolbar==4.2.0
django-filter==23.5
django-mail-templated==2.6.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
flake8==7.0.0
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
isodate==0.6.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.1.0
Markdown==3.5.1
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rpds-py==0.17.1
setuptools==68.2.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.2.1
wheel==0.41.2
yarl==1.9.4
zeep==4.2.1
